Do not cleanup unconfirmed paid customers (Closes #182)

+++ cmd/orchestrator/job/cleanup_unconfirmed_customer.go
@@ -31,6 +31,7 @@ func (j *Jobs) CleanupUnconfirmedCustomer(c *cron.Cron) {
 		// delete unconfirmed account
 		err := j.dbClient.Conn().
 			Model(&users).
+			Relation("Plan").
 			Join("INNER JOIN user_plans up ON id = up.user_id").
 			Where("up.activated_at IS NULL").
 			Select()
@@ -48,8 +49,8 @@ func (j *Jobs) CleanupUnconfirmedCustomer(c *cron.Cron) {
 				j.sendEmailReminder(user, 2)
 			}
 
-			// delete only user that did confirm their email since 30 days
-			if time.Until(user.CreatedAt.Add(maxUnconfirmedTime)) <= 0 {
+			// delete only unsubcribed user that did confirm their email since 30 days
+			if !user.Plan.IsSubscribed() && time.Until(user.CreatedAt.Add(maxUnconfirmedTime)) <= 0 {
 				if err := j.userService.DeleteUser(user.Email); err != nil {
 					j.logger.Sugar().Errorf("failed deleting user %s: %w", user.Email, err)
 				}
